+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+
+## 43.1.0 - 2019-06-19
 - [Feature] Add icons.json to OUI repo in prep for moving all icons into repo and automating a build script.
 - [Feature] Add ability to provide **SelectDropdown** with an `initialPlaceholder` value to use when a value is not yet selected ([#1176](https://github.com/optimizely/oui/pull/1176))
     - Also changed the `value` prop to no longer be required